moved monitor around
