+++ protocols/manage/simulation.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/BurntSushi/toml"
 	"github.com/dedis/onet/log"
-	"github.com/dedis/onet/monitor"
 	"github.com/dedis/onet/sda"
+	"github.com/dedis/onet/simul/monitor"
 )
 
 /*
